fix(lap-726): resolve comments
